requirements.txt
discord.py[voice]
gspread
pylint